Consistently use GCP XML API (#4207)
* Consistently use GCP XML API

* Use updated fake-gcs-server

* Review feedback
