+++ object_store/src/client/mod.rs
@@ -26,6 +26,12 @@ pub mod retry;
 #[cfg(any(feature = "aws", feature = "gcp", feature = "azure"))]
 pub mod token;
 
+#[cfg(any(feature = "aws", feature = "gcp", feature = "azure"))]
+pub mod header;
+
+#[cfg(any(feature = "aws", feature = "gcp"))]
+pub mod list;
+
 use std::collections::HashMap;
 use std::str::FromStr;
 use std::time::Duration;
